requirements.txt
pillow
psutil
matplotlib
pytest
